Implement a `is serial-dedup` method modifier
Respectively marked methods will then instantly return and do the
processing in the background. Calls that happen, while the task is still
running will cause a subsequent run once the current one is finished, but
always only a single one.

These semantics are useful for methods that process some kind of worklist.
Users of the method will want to add one or more items to the worklist and
then call the method. If multiple such additions and calls happen, while
the worklist is still being processed only a single subsequent run will
happen.