Fix PRs without a head URL
PRs miss a head ref when the underlying branch or repo has been deleted.
But we are already not relying on the head repo anymore but instead use
the "magical" refs containing the PR commits that GitHub automatically
creates in the base repo. Switch the last remnants of head-url over to
base-url and remove head-url completely.